+++ src/menu.h
@@ -21,12 +21,9 @@ enum menu_item_index {
   MENU_ITEM_REFRESH = 100,
   MENU_ITEM_LOCATION_HISTORY = 200,
   MENU_ITEM_TOGGLE_FAVORITE_LOCATION = 201,
-  MENU_ITEM_UNIT_SYSTEM = 202,
+  MENU_ITEM_CONFIGURE = 202,
   MENU_ITEM_ABOUT = 203,
   MENU_ITEM_QUIT = 300,
-  MENU_ITEM_UNIT_SYSTEM_STANDARD = 600,
-  MENU_ITEM_UNIT_SYSTEM_METRIC = 601,
-  MENU_ITEM_UNIT_SYSTEM_IMPERIAL = 602,
   MENU_ITEM_EMPTY_LOCATION_HISTORY = 800,
 };
 
@@ -50,7 +47,6 @@ private:
 
   std::shared_ptr<Icons> icons;
 
-  std::array<imenuex, 4> unit_system_items;
   std::array<imenuex, LocationHistoryProxy::max_size + 2>
       location_history_items;
   MenuItems items;
@@ -72,8 +68,6 @@ private:
   void update_item_texts();
 
   void update_location_history_items();
-
-  void update_unit_system_bullet();
 
   void open_menu();
 };