+++ src/ui.cc
@@ -199,15 +199,8 @@ void Ui::handle_menu_item_selected(int item_index) {
   } else if (item_index == MENU_ITEM_TOGGLE_FAVORITE_LOCATION) {
     SendEvent(event_handler, EVT_CUSTOM,
               CustomEvent::toggle_current_location_favorite, 0);
-  } else if (item_index == MENU_ITEM_UNIT_SYSTEM_STANDARD) {
-    SendEvent(event_handler, EVT_CUSTOM, CustomEvent::change_unit_system,
-              UnitSystem::standard);
-  } else if (item_index == MENU_ITEM_UNIT_SYSTEM_METRIC) {
-    SendEvent(event_handler, EVT_CUSTOM, CustomEvent::change_unit_system,
-              UnitSystem::metric);
-  } else if (item_index == MENU_ITEM_UNIT_SYSTEM_IMPERIAL) {
-    SendEvent(event_handler, EVT_CUSTOM, CustomEvent::change_unit_system,
-              UnitSystem::imperial);
+  } else if (item_index == MENU_ITEM_CONFIGURE) {
+    SendEvent(event_handler, EVT_CUSTOM, CustomEvent::open_config_editor, 0);
   } else if (item_index == MENU_ITEM_ABOUT) {
     SendEvent(event_handler, EVT_CUSTOM, CustomEvent::show_about_dialog, 0);
   } else if (MENU_ITEM_EMPTY_LOCATION_HISTORY < item_index and